requirements.txt
uvicorn==0.24.0.post1
fastapi==0.104.1
python-dotenv==1.0.0
pyxios==1.0.2
cryptography==38.0.4
pydantic==2.5.2
requests